API-example/requirements.txt
Django==1.8.4
django-bootstrap3==6.2.2
django-extensions==1.5.7
flake8==2.4.1
mccabe==0.3.1
pep8==1.6.2
pyflakes==1.0.0
geocoder==1.6.4
requests==2.7.0
psycopg2==2.6.1
djangorestframework==3.2.4
tweepy==3.4.0